requirements.txt
grad_cam==1.4.6
matplotlib==3.6.2
networkx==2.8.4
numpy==1.23.4
opencv_contrib_python==4.7.0.68
opencv_python==4.7.0.72
pandas==2.0.1
pillow==9.2.0
scikit_image==0.19.3
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
ttach==0.0.3
xlwt==1.3.0
